Fixall, yaml, loader, clean up yaml reader, better installer package
