+++ LICENSE
@@ -1,6 +1,6 @@
 BSD 2-Clause License
 
-Copyright (c) 2020, The CRTC
+Copyright (c) 2019-2020, The CRTC
 All rights reserved.
 
 Redistribution and use in source and binary forms, with or without
